+++ components/modal/Update-Bus-Readiness-Modal/update-bus-readiness.module.css
@@ -1,0 +1,214 @@
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1050;
+  animation: fadeIn 0.4s ease-out;
+}
+
+.title {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-family: 'Open Sans';
+  font-size: 24px;
+  font-weight: 800;
+  justify-content: center;
+  color: #ffffff;
+  background-color: #961c1e;
+  padding: 0.3rem 1.0rem;
+  margin-top: 10px;
+}
+
+.modal {
+  background: white;
+  border-radius: 8px;
+  max-width: 900px;
+  width: 100%;
+  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2);
+  animation: slideUp 0.3s ease-out;
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  background: #961c1e;
+  color: white;
+  padding: 16px 20px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-radius: 8px 8px 0 0;
+}
+
+.closeBtn {
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  color: white;
+  opacity: 0.8;
+  cursor: pointer;
+  transition: opacity 0.2s ease;
+}
+.closeBtn:hover {
+  opacity: 1;
+}
+
+.body {
+  padding: 20px;
+  background: white;
+  display: flex;
+  gap: 32px;
+}
+
+.leftCol, .rightCol {
+  flex: 1;
+  min-width: 0;
+}
+
+.section {
+  background: #f8f9fa;
+  border: 1px solid #e9ecef;
+  border-radius: 6px;
+  padding: 16px;
+  margin-bottom: 20px;
+}
+
+.sectionTitle {
+  font-size: 1rem;
+  font-weight: 600;
+  color: #495057;
+  border-bottom: 1px solid #dee2e6;
+  padding-bottom: 2px;
+  text-align: left;
+}
+
+.vehicleGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 12px 32px;
+  margin-bottom: 12px;
+}
+
+.vehicleCheck {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.label {
+  display: block;
+  font-size: 0.9rem;
+  font-weight: 500;
+  color: #495057;
+  margin-bottom: 6px;
+  text-align: left;
+}
+
+.input {
+  width: 100%;
+  padding: 10px 12px;
+  font-size: 0.95rem;
+  border: 1px solid #ced4da;
+  border-radius: 4px;
+  transition: border-color 0.2s ease;
+}
+
+.input:hover {
+  border-color: #961c1e;
+}
+
+.input:focus {
+  border-color: #961c1e;
+  outline: none;
+  box-shadow: 0 0 0 0.2rem rgba(150, 28, 30, 0.25);
+}
+
+.footer {
+  display: flex;
+  justify-content: flex-end;
+  gap: 10px;
+  padding: 16px 20px;
+  background: #f8f9fa;
+  border-radius: 0 0 8px 8px;
+}
+
+.cancelBtn,
+.createBtn {
+  padding: 8px 20px;
+  font-weight: 500;
+  border-radius: 4px;
+  border: 1px solid transparent;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.cancelBtn {
+  background: #961C1E;
+  border: 1px solid #FFB71F;
+  color: white;
+}
+.cancelBtn:hover {
+  background: #db1212;
+  border-color: #5a6268;
+  color: white;
+}
+
+.createBtn {
+  width: 160px;
+  height: 35px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  border-radius: 8px;
+  background-color: #961C1E;
+  color: #FFFFFF;
+  font-size: 14px;
+  transition: all 0.3s ease-in-out;
+  cursor: pointer;
+  border: none;
+}
+
+.createBtn:hover {
+  background-color: #7e191b; /* optional hover effect */
+}
+
+.createBtn:disabled {
+  background: #cccccc;
+  color: #888888;
+  cursor: not-allowed;
+  border: none;
+}
+
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes slideUp {
+  from {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* Only inside your modal */
+.overlay .form-check-input[type="checkbox"] {
+  accent-color: #961C1E;
+}
